Add EF logic

+++ ProjectService/Services/DeleteProject.cs
@@ -1,4 +1,4 @@
-﻿using Domain.Entities;
+﻿using Core.Exceptions;
 using Domain.Interfaces;
 using Services.Interfaces.Project_Interfaces;
 
@@ -16,7 +16,8 @@ public class DeleteProject : IDeleteProject
     public async Task<int> DeleteProjectAsync(int id)
     {
         var existingProject = await _storeProject.GetProjectByIdAsync(id);
-        // Валидация (если не найден, бросаем исключение);
+        if (existingProject == null)
+            throw new ValidationException("Project was not found in database", string.Empty);
 
         return await _storeProject.DeleteProjectAsync(existingProject);
     }
